Require capability token for video upload endpoints (issue #10)
/api/upload/video and /api/upload/video/signature were fully open —
anonymous Cloudinary uploads. They now require ?token= matching a
testimonial request_token or an offboarding survey/testimonial token,
the same tokens that gate the pages the recorder is embedded in.
Frontend passes the page token at both call sites.

Also fixes VideoUploadResponse.duration int->float (Cloudinary returns
fractional seconds), which was the root cause of the triaged
test_upload_video_success failure.

+++ cwc-platform/backend/app/schemas/testimonial.py
@@ -162,7 +162,7 @@ class VideoUploadResponse(BaseModel):
 
     url: str
     public_id: str
-    duration: int
+    duration: float  # Cloudinary reports fractional seconds
     thumbnail_url: Optional[str] = None
